Models folder was created and return types replaced with related struct.
