+++ server.go
@@ -2,177 +2,53 @@ package ilkbyte
 
 import (
 	"net/http"
-	"time"
+	"strconv"
+
+	"github.com/ilkbyte/api-go/models"
 )
 
-type Server struct {
-	Pagination struct {
-		CurrentPage int `json:"current_page"`
-		TotalPage   int `json:"total_page"`
-	} `json:"pagination"`
-	ServerList []struct {
-		BandwidthLimit int64  `json:"bandwidth_limit"`
-		BandwidthUsage int    `json:"bandwidth_usage"`
-		DeletedTime    string `json:"deleted_time,omitempty"`
-		Ipv4           string `json:"ipv4"`
-		Ipv6           string `json:"ipv6"`
-		Name           string `json:"name"`
-		Osapp          string `json:"osapp"`
-		Service        string `json:"service"`
-		Status         string `json:"status"`
-		CreatedTime    string `json:"created_time,omitempty"`
-	} `json:"server_list"`
-}
-
-type ServerConfig struct {
-	Application []struct {
-		Code   int    `json:"code"`
-		Name   string `json:"name"`
-		System string `json:"system"`
-	} `json:"application"`
-	OperatingSystem []struct {
-		Code    int    `json:"code"`
-		Name    string `json:"name"`
-		Version string `json:"version"`
-	} `json:"operating_system"`
-	Package []struct {
-		Code     int    `json:"code"`
-		Features string `json:"features"`
-		Name     string `json:"name"`
-		Price    string `json:"price"`
-	} `json:"package"`
-}
-
-type ServerCreate struct {
-	ServerInfo struct {
-		IPV4     string `json:"ipv4"`
-		IPV6     string `json:"ipv6"`
-		Name     string `json:"name"`
-		Osapp    string `json:"osapp"`
-		Password string `json:"password"`
-		Service  string `json:"service"`
-		Username string `json:"username"`
-	} `json:"server_info"`
-}
-
-type ServerDetails struct {
-	Service        string `json:"service"`
-	Status         string `json:"status"`
-	IPV4           string `json:"ipv4"`
-	IPV6           string `json:"ipv6"`
-	BandwidthLimit int64  `json:"bandwidth_limit"`
-	BandwidthUsage int64  `json:"bandwidth_usage"`
-}
-
-type ServerPower struct {
-	Status string `json:"status"`
-}
-
-type ServerIPList struct {
-	IPV4 []struct {
-		Address string `json:"address"`
-		Reverse string `json:"reverse"`
-		ACLList string `json:"acl_list"`
-		Usage   string `json:"usage"`
-	} `json:"ipv4"`
-	IPV6 []struct {
-		Address string `json:"address"`
-		Reverse string `json:"reverse"`
-		ACLList string `json:"acl_list"`
-		Usage   string `json:"usage"`
-	} `json:"ipv6"`
-}
-
-type ServerIPLogs []struct {
-	IPPrefix string `json:"ip_prefix"`
-	IsLog    bool   `json:"is_log"`
-	IsPerson bool   `json:"is_person"`
-	LogFile  string `json:"log_file"`
-	RuleIn   string `json:"rule_in"`
-	RuleOut  string `json:"rule_out"`
-	RuleType string `json:"rule_type"`
-}
-
-type Snapshot struct {
-	Snapshots []struct {
-		Name        string `json:"name"`
-		Current     bool   `json:"current"`
-		State       string `json:"state"`
-		Location    string `json:"location"`
-		Parent      string `json:"parent"`
-		Children    int    `json:"children"`
-		Descendants int    `json:"descendants"`
-		Metadata    bool   `json:"metadata"`
-		Params      struct {
-			IsCron   bool      `json:"is_cron"`
-			LastRun  time.Time `json:"last_run"`
-			NextRun  time.Time `json:"next_run"`
-			CronTime string    `json:"cron_time"`
-		} `json:"params"`
-		Date time.Time `json:"date"`
-	} `json:"snapshots"`
-}
-
-type Backup struct {
-	Amount string `json:"amount"`
-	Backup struct {
-		Name       string    `json:"name"`
-		FileSize   string    `json:"file_size"`
-		FileHash   string    `json:"file_hash"`
-		IsLocked   bool      `json:"is_locked"`
-		BackupTime time.Time `json:"backup_time"`
-	} `json:"backup"`
-}
-
-type BackupRestoreResp struct {
-	Name       string    `json:"name"`
-	FileSize   string    `json:"file_size"`
-	FileHash   string    `json:"file_hash"`
-	IsLocked   bool      `json:"is_locked"`
-	BackupTime time.Time `json:"backup_time"`
-}
-
-func (c *Client) GetAllServers() (*Response, error) {
+func (c *Client) GetAllServers(page int) (*models.Server, error) {
+	params := map[string]string{
+		"p": strconv.Itoa(page),
+	}
 	req, err := http.NewRequest("GET", c.BaseURL+"/server/list/all", nil)
 	if err != nil {
 		return nil, err
 	}
 
-	res := Response{
-		Data: &Server{},
-	}
-	if err := c.sendRequest(req, &res, nil); err != nil {
+	var res models.Server
+	if err := c.sendRequest(req, &res, params); err != nil {
 		return nil, err
 	}
 
 	return &res, nil
 }
 
-func (c *Client) GetActiveServers() (*Response, error) {
+func (c *Client) GetActiveServers(page int) (*models.Server, error) {
+	params := map[string]string{
+		"p": strconv.Itoa(page),
+	}
+
 	req, err := http.NewRequest("GET", c.BaseURL+"/server/list", nil)
 	if err != nil {
 		return nil, err
 	}
 
-	res := Response{
-		Data: &Server{},
-	}
-	if err := c.sendRequest(req, &res, nil); err != nil {
+	var res models.Server
+	if err := c.sendRequest(req, &res, params); err != nil {
 		return nil, err
 	}
 
 	return &res, nil
 }
 
-func (c *Client) GetServerConfig() (*Response, error) {
+func (c *Client) GetServerConfig() (*models.ServerConfig, error) {
 	req, err := http.NewRequest("GET", c.BaseURL+"/server/create", nil)
 	if err != nil {
 		return nil, err
 	}
 
-	res := Response{
-		Data: &ServerConfig{},
-	}
+	var res models.ServerConfig
 	if err := c.sendRequest(req, &res, nil); err != nil {
 		return nil, err
 	}
@@ -180,7 +56,7 @@ func (c *Client) GetServerConfig() (*Response, error) {
 	return &res, nil
 }
 
-func (c *Client) CreateServer(name, username, password, osid, appid, packageid, sshkey string) (*Response, error) {
+func (c *Client) CreateServer(name, username, password, osid, appid, packageid, sshkey string) (*models.ServerCreate, error) {
 	params := make(map[string]string)
 	params["name"] = name
 	params["username"] = username
@@ -195,9 +71,7 @@ func (c *Client) CreateServer(name, username, password, osid, appid, packageid, 
 		return nil, err
 	}
 
-	res := Response{
-		Data: &ServerCreate{},
-	}
+	var res models.ServerCreate
 	if err := c.sendRequest(req, &res, params); err != nil {
 		return nil, err
 	}
@@ -205,15 +79,13 @@ func (c *Client) CreateServer(name, username, password, osid, appid, packageid, 
 	return &res, nil
 }
 
-func (c *Client) GetServerDetails(serverName string) (*Response, error) {
+func (c *Client) GetServerDetails(serverName string) (*models.ServerDetails, error) {
 	req, err := http.NewRequest("GET", c.BaseURL+"/server/manage/"+serverName+"/show", nil)
 	if err != nil {
 		return nil, err
 	}
 
-	res := Response{
-		Data: &ServerDetails{},
-	}
+	var res models.ServerDetails
 	if err := c.sendRequest(req, &res, nil); err != nil {
 		return nil, err
 	}
@@ -221,7 +93,7 @@ func (c *Client) GetServerDetails(serverName string) (*Response, error) {
 	return &res, nil
 }
 
-func (c *Client) PowerServer(serverName, set string) (*Response, error) {
+func (c *Client) PowerServer(serverName, set string) (*models.ServerPower, error) {
 	params := make(map[string]string)
 	params["set"] = set
 
@@ -231,9 +103,7 @@ func (c *Client) PowerServer(serverName, set string) (*Response, error) {
 		return nil, err
 	}
 
-	res := Response{
-		Data: &ServerPower{},
-	}
+	var res models.ServerPower
 	if err := c.sendRequest(req, &res, params); err != nil {
 		return nil, err
 	}
@@ -241,15 +111,13 @@ func (c *Client) PowerServer(serverName, set string) (*Response, error) {
 	return &res, nil
 }
 
-func (c *Client) GetServerIPList(serverName string) (*Response, error) {
+func (c *Client) GetServerIPList(serverName string) (*models.ServerIPList, error) {
 	req, err := http.NewRequest("GET", c.BaseURL+"/server/manage/"+serverName+"/ip/list", nil)
 	if err != nil {
 		return nil, err
 	}
 
-	res := Response{
-		Data: &ServerIPList{},
-	}
+	var res models.ServerIPList
 	if err := c.sendRequest(req, &res, nil); err != nil {
 		return nil, err
 	}
@@ -257,13 +125,13 @@ func (c *Client) GetServerIPList(serverName string) (*Response, error) {
 	return &res, nil
 }
 
-func (c *Client) GetServerIPLogs(serverName string) (*Response, error) {
+func (c *Client) GetServerIPLogs(serverName string) (*models.ServerIPLogs, error) {
 	req, err := http.NewRequest("GET", c.BaseURL+"/server/manage/"+serverName+"/ip/logs", nil)
 	if err != nil {
 		return nil, err
 	}
 
-	res := Response{}
+	var res models.ServerIPLogs
 	if err := c.sendRequest(req, &res, nil); err != nil {
 		return nil, err
 	}
@@ -281,7 +149,7 @@ func (c *Client) ServerIPRdns(serverName, ip, rdns string) (*Response, error) {
 		return nil, err
 	}
 
-	res := Response{}
+	var res Response
 	if err := c.sendRequest(req, &res, params); err != nil {
 		return nil, err
 	}
@@ -289,15 +157,13 @@ func (c *Client) ServerIPRdns(serverName, ip, rdns string) (*Response, error) {
 	return &res, nil
 }
 
-func (c *Client) GetAllSnapshots(serverName string) (*Response, error) {
+func (c *Client) GetAllSnapshots(serverName string) (*models.Snapshot, error) {
 	req, err := http.NewRequest("GET", c.BaseURL+"/server/manage/"+serverName+"/snapshot", nil)
 	if err != nil {
 		return nil, err
 	}
 
-	res := Response{
-		Data: &Snapshot{},
-	}
+	var res models.Snapshot
 	if err := c.sendRequest(req, &res, nil); err != nil {
 		return nil, err
 	}
@@ -410,15 +276,13 @@ func (c *Client) DeleteCronSnapshot(serverName, snapshotName string) (*Response,
 	return &res, nil
 }
 
-func (c *Client) GetAllBackup(serverName string) (*Response, error) {
+func (c *Client) GetAllBackup(serverName string) (*models.Backup, error) {
 	req, err := http.NewRequest("GET", c.BaseURL+"/server/manage/"+serverName+"/backup", nil)
 	if err != nil {
 		return nil, err
 	}
 
-	res := Response{
-		Data: &Backup{},
-	}
+	var res models.Backup
 	if err := c.sendRequest(req, &res, nil); err != nil {
 		return nil, err
 	}
@@ -426,7 +290,7 @@ func (c *Client) GetAllBackup(serverName string) (*Response, error) {
 	return &res, nil
 }
 
-func (c *Client) BackupRestore(serverName, backupName string) (*Response, error) {
+func (c *Client) BackupRestore(serverName, backupName string) (*models.BackupRestoreResp, error) {
 	params := make(map[string]string)
 	params["backup_name"] = backupName
 
@@ -435,10 +299,7 @@ func (c *Client) BackupRestore(serverName, backupName string) (*Response, error)
 		return nil, err
 	}
 
-	res := Response{
-		Data: &BackupRestoreResp{},
-	}
-
+	var res models.BackupRestoreResp
 	if err := c.sendRequest(req, &res, params); err != nil {
 		return nil, err
 	}
